pyproject.toml
[project]
name = "peterental-vapi"
version = "0.1.0"
description = "Pete Rental VAPI with DuckDuckGo Search"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "loguru>=0.7.2",
    "python-multipart>=0.0.6",
    # LangChain Dependencies (Essential for DuckDuckGo search)
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.27",
    "langchain-ollama>=0.3.6",
    "duckduckgo-search>=8.1.1",
    # Additional utilities
    "rich>=14.1.0",
    "aiohttp>=3.9.0",
    "ddgs>=9.5.4",
    "beautifulsoup4>=4.13.5",
    "schedule>=1.2.2",
    "playwright>=1.54.0",
    "langchain-openai>=0.3.31",
    # Microsoft Calendar Integration
    "msal>=1.24.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

